# meme_backend/src/main/resources/openapi/api.yml
openapi: 3.0.1
info:
  title: Meme backend
  version: 0.0.1
...
servers:
  - url: http://localhost:8080/api
    description: Generated server url
paths:
...
  /v1/post/{id}:
    get:
      summary: Returns post
      tags:
        - "Post"
      operationId: getPost
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            description: Post id
            example: 1
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/PostGetFull'
    put:
      summary: Updates post selected by id
      security:
        - BearerAuth:
            - USER_EDIT
            - MODERATOR_EDIT
      tags:
        - "Post"
      operationId: updatePost
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 1
            description: Id of post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostPut'
        required: true
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/PostGetFull'
    post:
      summary: Publishes post
      security:
        - BearerAuth:
            - USER_ADD
      tags:
        - "Post"
      operationId: updatePostPublish
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 1
            description: Id of post
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/PostGetFull'
    delete:
      summary: Deletes post by given id
      security:
        - BearerAuth:
            - MODERATOR_DELETE
            - USER_DELETE
      tags:
        - "Post"
      operationId: deletePost
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            description: Post id
            example: 1
      responses:
        '200':
          description: OK
...
  /v1/accountrole/{id}:
    get:
      summary: Get role for given id
      security:
        - BearerAuth:
            - ADMIN_READ
      tags:
        - "AccountRole"
      operationId: getRole
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 1
            description: Account role id
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/AccountRoleGet'
    put:
      summary: Get role for given id
      security:
        - BearerAuth:
            - ADMIN_EDIT
      tags:
        - "AccountRole"
      operationId: putRole
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 1
            description: Account role id
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountRolePut'
        required: true
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/AccountRoleGet'
    delete:
      summary: Remove role for given id
      security:
        - BearerAuth:
            - ADMIN_DELETE
      tags:
        - "AccountRole"
      operationId: deleteRole
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 1
            description: Account role id
      responses:
        '200':
          description: OK
  /v1/accountrole/{id}/default:
    put:
      summary: Sets new default role
      security:
        - BearerAuth:
            - ADMIN_EDIT
      tags:
        - "AccountRole"
      operationId: putDefaultRole
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 1
            description: Account role id
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/AccountRoleGet'
...
  /v2/post:
    post:
      summary: Creates new post
      security:
        - BearerAuth:
            - USER_ADD
      tags:
        - "Post"
      operationId: createPublishedPost
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - files
                - title
              type: object
              properties:
                title:
                  description: Title of post
                  type: string
                visible:
                  description: Determines if post is visible publicly
                  type: boolean
                files:
                  description: Files of post
                  type: array
                  items:
                    type: string
                    format: binary
                description:
                  description: Descriptions of file
                  type: array
                  items:
                    type: string
                tags:
                  description: Tags for post
                  type: array
                  items:
                    type: integer
                    format: int64
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/PostGetFull'
  /v1/post:
    get:
      summary: Returns unpublished if any exists post
      security:
        - BearerAuth:
            - USER_READ
      tags:
        - "Post"
      operationId: getUnpublishedPost
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/PostGetFull'
    post:
      summary: Creates new unpublished post
      security:
        - BearerAuth:
            - USER_ADD
      tags:
        - "Post"
      operationId: createUnpublishedPost
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostPost'
        required: true
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/PostGetFull'
...
  /v1/post/all:
    post:
      summary: Returns paginated list of posts
      tags:
        - "Post"
      operationId: getAllPosts
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/PostPageGet'
...
  /v1/accountrole/all:
    get:
      summary: Returns all account roles
      security:
        - BearerAuth:
            - ADMIN_READ
      tags:
        - "AccountRole"
      operationId: getRoles
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountRoleGet'
    post:
      summary: Creates new account role
      security:
        - BearerAuth:
            - ADMIN_ADD
      tags:
        - "AccountRole"
      operationId: postRole
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountRolePost'
        required: true
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/AccountRoleGet'
...
  /v1/comment/{id}/commentstatistic:
    get:
      summary: Returns comment statistics
      tags:
        - "CommentStatistic"
      operationId: getCommentStatistic
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            description: Comment id
            example: 1
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/CommentStatisticGet'
...
  /v1/comment/{id}:
    delete:
      summary: Deletes comment by given id
      security:
        - BearerAuth:
            - MODERATOR_DELETE
            - USER_DELETE
      tags:
        - "Comment"
      operationId: deleteComment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
components:
  schemas:
    AccountGetFull:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Id of Account
          example: 2
        name:
          type: string
          description: Name of account
          example: BigBen
        email:
          type: string
          format: email
          description: Email of account
          example: BigBen@gmail.com
        creationDate:
          type: string
          format: date
          description: CreationDate of account
          example: 01.01.2000
        profilePhotoId:
          type: integer
          format: int64
          description: Id of profile photo
          example: 1
        role:
          $ref: '#/components/schemas/AccountRoleGet'
        enabled:
          type: boolean
          description: Determines if account is enabled
          example: true
        banned:
          type: boolean
          description: Determines if account is banned
          example: true
    AccountGetShort:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Id of Account
          example: 2
        name:
          type: string
          description: Name of account
          example: BigBen
        creationDate:
          type: string
          format: date
          description: CreationDate of account
          example: 01.01.2000
        profilePhotoId:
          type: integer
          format: int64
          description: Id of profile photo
          example: 1
    AccountPermissionGet:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Id of permission
          example: 1
        permission:
          type: string
          description: Name of permission
          example: USER_READ
    CommentGet:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Id of comment
          example: 1
        account:
          $ref: '#/components/schemas/AccountGetShort'
        replyToId:
          type: integer
          format: int64
          description: Id of comment this comment is reply to
          example: 2
        comment:
          type: string
          description: Comment value
          example: Ha ha funny
        commentStatistics:
          $ref: '#/components/schemas/CommentStatisticGet'
    CommentStatisticGet:
      type: object
      properties:
        commentId:
          type: integer
          format: int64
          description: Id of comment
          example: 2
        upVoteCount:
          type: integer
          format: int64
          description: Ammount of positive votes this comment have
          example: 2
        downVoteCount:
          type: integer
          format: int64
          description: Ammount of negative votes this comment have
          example: 2
    PostGetFull:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Id of post
          example: 1
        title:
          type: string
          description: Title of post
          example: Ha ha funny
        creationDate:
          type: string
          format: date
          description: Creation date of post
          example: 01.01.2000
        visible:
          type: boolean
          description: Determines if post is publicly visible
          example: true
        filesId:
          type: array
          description: Array of post object ids
          items:
            type: integer
            format: int64
            description: Post object id
            example: 1
        tags:
          type: array
          description: Array of tags
          items:
            $ref: '#/components/schemas/TagGet'
        comments:
          type: array
          description: Array of comments
          items:
            $ref: '#/components/schemas/CommentGet'
        account:
          $ref: '#/components/schemas/AccountGetFull'
        postStatistics:
          $ref: '#/components/schemas/PostStatisticGet'
    PostStatisticGet:
      type: object
      properties:
        postId:
          type: integer
          format: int64
          description: Post statistic id
          example: 1
        seenCount:
          type: integer
          format: int64
          description: Ammount of times this post was seen by unique user
          example: 1
        upVoteCount:
          type: integer
          format: int64
          description: Ammount of positive votes this post have
          example: 2
        downVoteCount:
          type: integer
          format: int64
          description: Ammount of negative votes this post have
          example: 2
        favoriteCount:
          type: integer
          format: int64
          description: Ammount of favorites this post have
          example: 2
    TagGet:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Id of tag
          example: 1
        tag:
          type: string
          description: Value of tag
          example: Cat
    PostPut:
      type: object
      required:
        - title
        - tags
        - visible
        - order
      properties:
        title:
          type: string
          description: Title of post
          example: Ha ha funny
        tags:
          type: array
          description: List of tag ids
          items:
            type: integer
            format: int64
            example: 1
            description: Id of tag
        visible:
          type: boolean
          example: true
          description: Determines if post is visible in public
        order:
          type: object
          description: Map of PostObject Ids and weights determining ordering of post objects returned
          additionalProperties:
            type: integer
            format: int64
            description: Weight of PostObject, lower values are more important
...
    AccountRolePut:
      type: object
      required:
        - permissions
        - role
      properties:
        permissions:
          type: array
          description: Array of permission ids
          items:
            type: integer
            format: int64
            description: Permission id
            example: 1
        role:
          type: string
          description: Role name
          example: USER
    AccountRoleGet:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Account role id
          example: 1
        role:
          type: string
          description: Role name
          example: USER
        permissions:
          type: array
          description: Array of permissions
          items:
            $ref: '#/components/schemas/AccountPermissionGet'
    PostPost:
      type: object
      required:
        - title
        - visible
      properties:
        title:
          type: string
          description: Title of post
          example: Ha ha funny
        tags:
          type: array
          description: Array of tags id
          items:
            type: integer
            format: int64
            description: Tag id
            example: 1
        visible:
          type: boolean
          description: Determines if post is visible publicly
          example: true
...
    DateRange:
      type: object
      nullable: true
      required:
        - min
        - max
      properties:
        min:
          type: string
          format: date
          description: Minimal date to consider
          example: 01.01.2000
        max:
          type: string
          format: date
          description: Maximal date to consider
          example: 01.01.2000
    LongRange:
      nullable: true
      type: object
      required:
        - min
        - max
      properties:
        min:
          type: integer
          format: int64
          description: Minimal value to consider
          example: 1
        max:
          type: integer
          format: int64
          description: Maximal value to consider
          example: 5
    PostFilter:
      type: object
      properties:
        creationDate:
          $ref: '#/components/schemas/DateRange'
        comments:
          $ref: '#/components/schemas/LongRange'
        upVotes:
          $ref: '#/components/schemas/LongRange'
        downVotes:
          $ref: '#/components/schemas/LongRange'
        favorites:
          $ref: '#/components/schemas/LongRange'
        tags:
          type: array
          nullable: true
          description: Array of tag ids to filter by
          items:
            type: integer
            format: int64
            description: Tag id
            example: 1
        title:
          type: string
          nullable: true
          description: Title of post to filter by
          example: Ha ha
    PostPagingAndSortingRequest:
      type: object
      required:
        - page
        - size
      properties:
        page:
          type: integer
          format: int32
          description: Page to display
          example: 1
        size:
          type: integer
          format: int32
          description: Size of page to display
          example: 1
        sortBy:
          type: string
          description: Sort by field
          example: id
        sortDirection:
          type: string
          description: Sort direction
          example: asc
    PostRequest:
      type: object
      properties:
        pagingAndSorting:
          $ref: '#/components/schemas/PostPagingAndSortingRequest'
        filters:
          $ref: '#/components/schemas/PostFilter'
    PostGetShort:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Id of post
          example: 1
        title:
          type: string
          description: Title of post
          example: Ha ha funny
        firstFileId:
          type: integer
          format: int64
          description: First post object id
          example: 1
        account:
          $ref: '#/components/schemas/AccountGetFull'
        postStatistics:
          $ref: '#/components/schemas/PostStatisticGet'
    PostPageGet:
      type: object
      properties:
        posts:
          type: array
          description: Array of posts
          items:
            $ref: '#/components/schemas/PostGetShort'
        totalNumberOfPages:
          type: integer
          format: int32
          description: Total number of pages with current filter applied
          example: 1
        totalNumberOfElements:
          type: integer
          format: int64
          description: Total number of elements that can be queried
          example: 1
        filters:
          $ref: '#/components/schemas/PostFilter'
...
    AccountRolePost:
      type: object
      required:
        - permissions
        - role
      properties:
        permissions:
          type: array
          description: Array of permission ids
          items:
            type: integer
            format: int64
        role:
          type: string
          description: Account role
          example: User
...